Miscellaneous fixes:
Reference elements: when checking if an element's identifier is unique, exclude elements that contain or are contained by the element label. This bug was causing us to add a reference element in situations where it isn't necessary.
Labels: for labels for non-input elements, but make them highConfidence = false. This allows us to pick up the label from dropdowns constructed with divs and clicks on divs that contain a switch-label pair
Descriptors: ignore innerText if only contains whitespace (it is going to be trimmed on the backend so we need to ignore it)